Move all Wazi Deploy Application Manifest actions into utilities class
