+++ Pipeline/PackageBuildOutputs/PackageBuildOutputs.groovy
@@ -11,7 +11,6 @@ import groovy.cli.commons.*
 import java.nio.file.*
 import static java.nio.file.StandardCopyOption.*
 import com.ibm.jzos.ZFile;
-import groovy.yaml.YamlBuilder
 
 /************************************************************************************
  * This script creates a simplified package with the outputs generated from a DBB build
@@ -43,13 +42,16 @@ import groovy.yaml.YamlBuilder
  *  - Added support for DeployableArtifact, to manage the correct stacking
  *    of duplicates artifacts
  * 
- * Version 6 - 2024-02
+ * Version 6 - 2024-03
  *  - Added support to write IBM Wazi Deploy Application Manifest file
  *      
  ************************************************************************************/
 
 // start create & publish package
 @Field Properties props = null
+def scriptDir = new File(getClass().protectionDomain.codeSource.location.path).parent
+@Field def wdManifestGeneratorUtilities = loadScript(new File("${scriptDir}/utilities/WaziDeployManifestGenerator.groovy"))
+
 props = parseInput(args)
 
 def startTime = new Date()
@@ -76,7 +78,7 @@ Map<DeployableArtifact, Map> buildOutputsMap = new HashMap<DeployableArtifact, M
 def String tarFileLabel = "Default"
 
 // Object to store scm information for Wazi Deploy Application Manifest file
-@Field ScmInfo scmInfo = new ScmInfo()
+HashMap<String,String> scmInfo = new HashMap<String, String>()
 
 // iterate over all build reports to obtain build output
 props.buildReportOrder.each { buildReportFile ->
@@ -229,18 +231,19 @@ props.buildReportOrder.each { buildReportFile ->
 		}
 	}
 
-	// generate scmInfo record
-	if (props.buildReportOrder.size() == 1) {
-		scmInfo.type = "git"
-		gitUrl = retrieveBuildResultProperty (buildResultPropertiesRecord, "giturl")
-		if (gitUrl) scmInfo.uri = gitUrl
-		gitHash = retrieveBuildResultProperty (buildResultPropertiesRecord, "githash")
-		if (gitHash) scmInfo.shortCommit = gitHash
-		scmInfo.branch = (props.branch) ? props.branch : "UNDEFINED"
-
-	} else {
-		scmInfo.shortCommit = "multipleBuildReports"
-		scmInfo.uri = "multipleBuildReports"
+	// generate scmInfo for Wazi Deploy Application Manifest file
+	if (props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()) {
+		if (props.buildReportOrder.size() == 1) {
+			scmInfo.put("type", "git")
+			gitUrl = retrieveBuildResultProperty (buildResultPropertiesRecord, "giturl")
+			if (gitUrl) scmInfo.put("uri", gitUrl)
+			gitHash = retrieveBuildResultProperty (buildResultPropertiesRecord, "githash")
+			if (gitHash) scmInfo.put("shortCommit", gitHash)
+			scmInfo.put("branch", props.branch)
+		} else {
+			scmInfo.put("shortCommit", "multipleBuildReports")
+			scmInfo.put("uri", "multipleBuildReports")
+		}
 	}
 
 }
@@ -254,8 +257,11 @@ if (buildOutputsMap.size() == 0) {
 } else {
 
 	// Local variables
-	ApplicationManifest applicationManifest // Wazi Deploy Application Manifest
-	if (props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()) applicationManifest = initializeApplicationManifest()
+	// Initialize Wazi Deploy Manifest Generator
+	if (props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()) {
+		wdManifestGeneratorUtilities.initWaziDeployManifestGenerator(props)// Wazi Deploy Application Manifest
+		wdManifestGeneratorUtilities.setScmInfo(scmInfo)
+	}
 	def String tarFileName = (props.tarFileName) ? props.tarFileName  : "${tarFileLabel}.tar"
 	def tarFile = "$props.workDir/${tarFileName}"
 
@@ -323,8 +329,8 @@ if (buildOutputsMap.size() == 0) {
 					}
 
 					// Append record to Wazi Deploy Application Manifest
-					if (applicationManifest && props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()) {
-						appendArtifactToAppManifest(deployableArtifact, record, propertiesRecord, applicationManifest)
+					if (wdManifestGeneratorUtilities && props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()) {
+						wdManifestGeneratorUtilities.appendArtifactToAppManifest(deployableArtifact, record, propertiesRecord)
 					}
 
 				} else {
@@ -338,9 +344,9 @@ if (buildOutputsMap.size() == 0) {
 		}
 	}
 
-	if (applicationManifest && props.generateWaziDeployAppManifest && !props.error) {
+	if (wdManifestGeneratorUtilities && props.generateWaziDeployAppManifest && !props.error) {
 		// print application manifest
-		writeApplicationManifest(new File("$tempLoadDir/waziDeployManifest.yaml"), applicationManifest)
+		wdManifestGeneratorUtilities.writeApplicationManifest(new File("$tempLoadDir/waziDeployManifest.yaml"), props.verbose)
 	}
 
 	if (!props.error) {
@@ -445,7 +451,6 @@ if (buildOutputsMap.size() == 0) {
 		def repo = props.get('artifactRepository.repo') as String
 
 		//Call the artifactRepositoryHelpers to publish the tar file
-		def scriptDir = new File(getClass().protectionDomain.codeSource.location.path).parent
 		File artifactRepoHelpersFile = new File("$scriptDir/ArtifactRepositoryHelpers.groovy")
 		Class artifactRepositoryHelpersClass = new GroovyClassLoader(getClass().getClassLoader()).parseClass(artifactRepoHelpersFile)
 		GroovyObject artifactRepositoryHelpers = (GroovyObject) artifactRepositoryHelpersClass.newInstance()
@@ -659,7 +664,7 @@ def parseInput(String[] cliArgs){
 			assert props.get("artifactRepository.password") : "Missing artifact repository Password"
 		}
 
-		// TODO wazi deploy options
+		// assess required options to generate Wazi Deploy application manifest
 		if(props.generateWaziDeployAppManifest && props.generateWaziDeployAppManifest.toBoolean()){
 			assert props.get("branch") : "Generating the Wazi Deploy Application Missing requires that the branch is passed to PackageBuildOutputs. See parameter --branch."
 			assert props.get("addExtension") : "Generating the Wazi Deploy Application Missing requires that to activate adding the deployType as file extensions. See parameter --addExtension."
@@ -724,151 +729,7 @@ class DeployableArtifact {
 	}
 }
 
-/**
- * Wazi Deploy Application Manifest Classes and Helpers
- */
-
-class ApplicationManifest {
-	String apiVersion = "wazideploy.ibm.com/v1"
-	String kind = "ManifestState"
-	Metadata metadata
-	ArrayList<Artifact> artifacts
-}
-
-class Metadata {
-	String name
-	String description
-	String version
-	Annotations annotations
-}
-
-class Annotations {
-	String creationTimestamp
-	ScmInfo scmInfo
-	PackageInfo packageInfo
-}
-
-class ScmInfo {
-	String type
-	String uri
-	String branch
-	String shortCommit
-}
-
-class PackageInfo {
-	String name
-	String description
-	Properties properties
-	String uri
-	String type
-}
-
-class Artifact {
-	String name
-	String description
-	ArrayList<ElementProperty> properties
-	String type
-	String hash
-}
-
-class ElementProperty {
-	String key
-	String value
-}
-
-/**
- *
- */ 
-def appendArtifactToAppManifest(DeployableArtifact deployableArtifact, Record record, PropertiesRecord propertiesRecord, ApplicationManifest applicationManifest){
-	Artifact artifact = new Artifact()
-	artifact.name = deployableArtifact.file
-	gitHashInfo = retrieveBuildResultProperty (propertiesRecord, "githash")
-	artifact.hash =  (gitHashInfo) ? gitHashInfo : "UNDEFINED"
-	artifact.description = (record.file) ? record.file : deployableArtifact.file
-	if (propertiesRecord) {
-		ArrayList<ElementProperty> artifactProperties = new ArrayList()
-		["githash", "giturl"].each {property ->
-			ElementProperty artifactProperty = new ElementProperty()
-			propValue =  retrieveBuildResultProperty (propertiesRecord, property)
-			if (propValue) {
-				artifactProperty.key = property
-				artifactProperty.value = propValue
-				artifactProperties.add(artifactProperty)
-			}
-		}
-		if (artifactProperties.size() > 0 ) {
-			artifact.properties = artifactProperties
-		}
-	}
-	artifact.type =deployableArtifact.deployType
-
-	// adding artifact into applicationManifest
-	applicationManifest.artifacts.add(artifact)
-}
-
-/**
- * Write an Wazi Deploy Manifest  a YAML file
- */
-def writeApplicationManifest(File yamlFile, ApplicationManifest applicationManifest){
-	println("** Generate Wazi Deploy Application Manifest file to $yamlFile")
-	def yamlBuilder = new YamlBuilder()
-
-	yamlBuilder {
-		apiVersion applicationManifest.apiVersion
-		kind applicationManifest.kind
-		metadata applicationManifest.metadata
-		artifacts  applicationManifest.artifacts
-	}
-
-	if (props.verbose && props.verbose.toBoolean()) {
-		println yamlBuilder.toString()
-	}
-
-	// write file
-	yamlFile.withWriter() { writer ->
-		writer.write(yamlBuilder.toString())
-	}
-}
-
-/**
- * Initialize application manifest file 
- */
-
-def initializeApplicationManifest() {
-	ApplicationManifest applicationManifest = new ApplicationManifest()
-	ArrayList<Artifact> artifacts = new ArrayList()
-	applicationManifest.artifacts = artifacts
-
-	// Metadata
-	Metadata metadata = new Metadata()
-
-	applicationManifest.metadata = metadata
-
-	// Annotations
-	Annotations annotations = new Annotations()
-	annotations.creationTimestamp = props.startTime
-	metadata.annotations = annotations
-	if (props.application) {
-		metadata.description = props.application
-		metadata.name = props.application
-	} else {
-		metadata.description = "UNDEFINED"
-		metadata.name = "UNDEFINED"
-	}
-
-	// Fields to write Application Manifest file
-
-	PackageInfo packageInfo = new PackageInfo()
-
-	if (scmInfo) annotations.scmInfo = scmInfo
-	metadata.version = (props.versionName) ? props.versionName : props.startTime
-	if (props.application) metadata.name = props.application
-
-	return applicationManifest
-}
-
 def retrieveBuildResultProperty(PropertiesRecord buildResultPropertiesRecord, String propertyName) {
-
 
 	if (buildResultPropertiesRecord!=null) {
 		buildResultProperties = buildResultPropertiesRecord.getProperties()
@@ -884,5 +745,4 @@ def retrieveBuildResultProperty(PropertiesRecord buildResultPropertiesRecord, St
 			return null
 		}
 	}
-
 }
